refactor: separate logic for adicionar page
